Refactored HintProviderService interface
