+++ src/HuntAndPeck/Extensions/HintProviderExtentions.cs
@@ -1,0 +1,112 @@
+﻿using HuntAndPeck.Models;
+using HuntAndPeck.NativeMethods;
+using HuntAndPeck.Services.Interfaces;
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using UIAutomationClient;
+
+namespace HuntAndPeck.Extensions
+{
+    public static class HintProviderExtentions
+    {
+        private static readonly IUIAutomation _automation = new CUIAutomation();
+
+        public static HintSession EnumHints<THint>(this IHintProviderService<THint> hintProviderService)
+            where THint : Hint
+        {
+            var foregroundWindow = User32.GetForegroundWindow();
+            if (foregroundWindow == IntPtr.Zero)
+            {
+                return null;
+            }
+            return EnumHints(hintProviderService, foregroundWindow);
+        }
+    
+        public static HintSession EnumHints<THint>(this IHintProviderService<THint> hintProviderService, IntPtr hWnd)
+            where THint : Hint
+        {
+            Stopwatch sw = new Stopwatch();
+            sw.Start();
+            var session = EnumWindowHints(hWnd, hintProviderService.CreateHint);
+            sw.Stop();
+
+            Debug.WriteLine("Enumeration of hints took {0} ms", sw.ElapsedMilliseconds);
+            return session;
+        }
+        /// <summary>
+        /// Enumerates all the hints from the given window
+        /// </summary>
+        /// <param name="hWnd">The window to get hints from</param>
+        /// <param name="hintFactory">The factory to use to create each hint in the session</param>
+        /// <returns>A hint session</returns>
+        private static HintSession EnumWindowHints(IntPtr hWnd, Func<IntPtr, Rect, IUIAutomationElement, Hint> hintFactory)
+        {
+            var result = new List<Hint>();
+            var elements = EnumElements(hWnd);
+
+            // Window bounds
+            var rawWindowBounds = new RECT();
+            User32.GetWindowRect(hWnd, ref rawWindowBounds);
+            Rect windowBounds = rawWindowBounds;
+
+            foreach (var element in elements)
+            {
+                var boundingRectObject = element.CurrentBoundingRectangle;
+                if ((boundingRectObject.right > boundingRectObject.left) && (boundingRectObject.bottom > boundingRectObject.top))
+                {
+                    var niceRect = new Rect(new Point(boundingRectObject.left, boundingRectObject.top), new Point(boundingRectObject.right, boundingRectObject.bottom));
+                    // Convert the bounding rect to logical coords
+                    var logicalRect = niceRect.PhysicalToLogicalRect(hWnd);
+                    if (!logicalRect.IsEmpty)
+                    {
+                        var windowCoords = niceRect.ScreenToWindowCoordinates(windowBounds);
+                        var hint = hintFactory(hWnd, windowCoords, element);
+                        if (hint != null)
+                        {
+                            result.Add(hint);
+                        }
+                    }
+                }
+            }
+
+            return new HintSession
+            {
+                Hints = result,
+                OwningWindow = hWnd,
+                OwningWindowBounds = windowBounds,
+            };
+        }
+
+        /// <summary>
+        /// Enumerates the automation elements from the given window
+        /// </summary>
+        /// <param name="hWnd">The window handle</param>
+        /// <returns>All of the automation elements found</returns>
+        private static List<IUIAutomationElement> EnumElements(IntPtr hWnd)
+        {
+            var result = new List<IUIAutomationElement>();
+            var automationElement = _automation.ElementFromHandle(hWnd);
+
+            var conditionControlView = _automation.ControlViewCondition;
+            var conditionEnabled = _automation.CreatePropertyCondition(UIA_PropertyIds.UIA_IsEnabledPropertyId, true);
+            var enabledControlCondition = _automation.CreateAndCondition(conditionControlView, conditionEnabled);
+
+            var conditionOnScreen = _automation.CreatePropertyCondition(UIA_PropertyIds.UIA_IsOffscreenPropertyId, false);
+            var condition = _automation.CreateAndCondition(enabledControlCondition, conditionOnScreen);
+
+            var elementArray = automationElement.FindAll(TreeScope.TreeScope_Descendants, condition);
+            for (var i = 0; i < elementArray.Length; ++i)
+            {
+                result.Add(elementArray.GetElement(i));
+            }
+
+            return result;
+        }
+
+    }
+}
