+++ src/HuntAndPeck/HuntAndPeck.csproj
@@ -85,6 +85,7 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="Extensions\HintProviderExtentions.cs" />
     <Compile Include="Extensions\IEnumerableTExtensions.cs" />
     <Compile Include="Extensions\RectExtensions.cs" />
     <Compile Include="Models\DebugHint.cs" />
@@ -95,8 +96,9 @@
     <Compile Include="Models\UiAutomationInvokeHint.cs" />
     <Compile Include="Models\UiAutomationSelectHint.cs" />
     <Compile Include="Models\UiAutomationToggleHint.cs" />
+    <Compile Include="Services\DebugHintProviderService.cs" />
+    <Compile Include="Services\GeneralHintProviderService.cs" />
     <Compile Include="Services\HintLabelService.cs" />
-    <Compile Include="Services\Interfaces\IDebugHintProviderService.cs" />
     <Compile Include="Services\Interfaces\IHintLabelService.cs" />
     <Compile Include="Services\Interfaces\IHintProviderService.cs" />
     <Compile Include="Services\Interfaces\IKeyListenerService.cs" />
@@ -104,7 +106,6 @@
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Services\UiAutomationPatternIds.cs" />
-    <Compile Include="Services\UiAutomationHintProviderService.cs" />
     <Compile Include="SingleLaunchMutex.cs" />
     <Compile Include="ViewModels\DebugHintViewModel.cs" />
     <Compile Include="ViewModels\DebugOverlayViewModel.cs" />
