--- src/HuntAndPeck/Services/Interfaces/IDebugHintProviderService.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using HuntAndPeck.Models;
-
-namespace HuntAndPeck.Services.Interfaces
-{
-    public interface IDebugHintProviderService
-    {
-        HintSession EnumDebugHints();
-        HintSession EnumDebugHints(IntPtr hWnd);
-    }
-}
